move OCS related steps to own context
